Change scope of counter variable to allow its use outside loop
When selecting a random plot to be the target of global warming a counter is used to only select a set number of map plots to be candidates. The counter variable is declared within the for loop but is referenced after the loop completes.

Declaring the variable prior to the loop allows the variable to be used in a wider scope.

Resolves: AC4-32