Change scope of counter variable to allow its use outside loop
In CvUnitAI::AI_retreatToCity, when determining whether a unit should push the mission to retreat to a city a counter is used to counter is used to determine the type of movement required. The counter variable is declared within the for loop but is referenced after the loop completes. In order to resolve this the counter scope has been expanded.

Resolves: AC4-33