Limit unit movement range
Unit movement may be restricted to one of 4 levels defined on the unit info:
1 - Home city plots
2 - Player territory
3 - Range from home city
4 - Unlimited

A unit that is designated as bound to the Home city plots will never be able to leave those irrespective of any player modifiers or promotions.

The base range from the home city is defined in XML as a GlobalDefine and is scaled by the world size.

Units that find themselves outside of their designated range through events such as the loss of a city, etc., may always move towards their newly permitted plots.

AC4-41
AC4-30