Allow units to require civics
Units can have Civic AND or OR prereqs

If the player is not running the civic(s) then the unit will refuse to fight/build/found and becomes home bound so will not be able to move further away from their home city. If the player starts running the civic(s) again then the unit returns back to normal.

AC4-53
AC4-30