Rework existing BuildingClass prereqs on buildings
There are 2 OOB prereqs for buildings related to BuildingClass objects. These are a minimum number of these classes in player cities across the civ and a set of BuildingClasses that must exist in the city before the building can be constructed.

The naming of these in the XML files is not consistent with the new building prereqs added so these have been updated to make it clearer what they represent.

The implementation with the CvBuildingInfo class has also been updated to bring them in line with the new prereqs. In the case of the set of AND BuildingClasses these work in a similar way to the new OR BuildingClass prereq where the building is disabled if the prereq ceases to be met.

AC4-69
AC4-30