Make XP values fractional
It is now possible to give a unit a fractional amount of XP allowing for more granular XP gain and also opening up the possibility for XP gain to be made in more scenarios without overly impacting the game balance. An example would be allowing healing units to gain XP when they provide healing to a unit, or scouts on exploration. This means that non-combat units can gain promotions realistically.

AC4-76