Add Occasional Promotions
Added the concept of occasional promotions where there is a chance that a unit will gain a promotion following a battle. Any promotion gained must be valid for the unit and relevant to the combat, e.g., when fighting in woods you may get the Woodsman line or when fighting across a river you may Amphibious.

AC4-96