Add the Natural Wonders mod from PlatyPing
Implement the mod in the dll rather than the original Python which allows the AI to evaluate these plots

This adds unique features to the map with each feature having a hidden building associated with it that provides benefits to the city that has the feature in its working area. There are various benefits for each feature, but all provide some Great Person points for one kind of GP. The features also have a random yield values, the min and max of which are set as global defines.

AC4-102