Add star signs mod
Add the concept of star signs where events occur on a regular basis which can have both negative or positive impacts on the empire of all players. The scale and scope of the event may be mitigated by buildings, civics, traits and specialists. The effect may happen in a single city or across more than one city and the scale of the effect is calculated for each city individually.

AC4-112
AC4-30