Adjusted CvTeam::shareCounters to prevent lose of circumnavigation bonus.
There is no perfect way to share extra moves when teams merge - because
the source of extra moves is not tracked, and so we can't tell if the
two teams have the same bonuses, or independent bonuses.

I've decided to just take the maximum value of the two; which for most
cases will be correct.

Commit comment from K-Mod git repo

AC4-129