Fixed air intercepts conflicting with combat in simultaneous turns.
Air intercept battles previous did not check if the intercepting unit
was already fighting. This could cause bad things to happen.

Also, air intercept battles no longer check if their is combat on the
intercepting unit's plot. This means that the air combot & intercept can
happen while the plot is being attacked (as long as the intercepting
unit itself is not fighting). This is a somewhat experimental change...
I can't think of any problems with it; but maybe there is something I
haven't thought of.

Commit comment from K-Mod git repo

AC4-129