"RiverSideYield" for improvements is now given for touching a river.
Originally this bonus required the improvement to be on the side of a river, like a watermill. Now the bonus is applied to all river tiles, like it is for bare tiles. ("RiverSide" is now a misnomer, but I don't think it's worth changing.)

Commit comment from K-Mod git repo

AC4-129