Changed some corporation maintenance.
Increased maintenance for Civilized Jewelers. Decreased maintenance for Std. Ethanol and Aluminium Co.

The fact that Civ. Jewelers is unlocked so early makes it a powerful way to boost a civilization's economy even ignoring the culture provided. Increasing the maintenance doesn't seem like a great 'fix'; but I do think Civ Jewelers was a bit too powerful.

The other two corporations are in the opposite situation. They are unlocked late in the game, and their bonuses are situational anyway.

Commit comment from K-Mod git repo

AC4-129