mproved clarity and consistency in CvPlayerAI::AI_counterPropose
Reduced the scope of temporary variables, and changed the usage of
'weight' and 'value' to be consistent.

'value' means how much the item would add to the trade deal; whereas
'weight' is an arbitrary scale used to decide which items we should
prefer to trade.

I hope that these changes will reduce the likelihood of mistakes.

Commit comment from K-Mod git repo

AC4-129