AI_updateCommerceWeights now better understands focused espioange.
Having very high values on the espionage weight slider are now taken to mean that you don't want to spend espionage on the other teams. This reduces the chance of the AI choosing to assign spies inapproriately.

Previously, having a non-zero weight against a team was taken to mean that the player wanted to espionage points against that team - and thus espionage was more highly valued if the points were low.

Now that is only the case if the weight is above a certain threshold. The threshold is based on total weight, and number of teams etc.

Commit comment from K-Mod git repo

AC4-129