Minor tweaks to CvCityAI::AI_buildingValue
Reduced building value of free-tech. (The evaluation for this is still just a very poor guess; but it is very difficult to evaluate it properly. A true evaluation would require details which we currently do not take into account.)

Reduced the value of espionage commerce multipliers.

Commit comment from K-Mod git repo

AC4-129