Increased trade culture rate.
Instead of 1% per culture level, it's now the average of that and the max rate. i.e. percentage = (current level + max level)/2;

(This should probably be changed so that it doesn't depend on how many culture levels are in the game; otherwise it might get weird in some mods.)

Commit comment from K-Mod git repo

AC4-129