Rewrote CvPlayerAI::AI_bestTech to better consider tech pathways.
Previously the AI would choose to beeline high-value techs within the depth limit without considering the prerequisites at all. This sometime led to poor choices; e.g. getting stuck on very high cost prereqs, or researching prereqs which would obsolete important bonuses.

In the new system, the AI only ever chooses techs that they can research; but adds value based on which techs follow on.

NOTE: this is a totally new and system which will requires some changes and balance. Initial testing suggests that in some cases it is actually _worse_ than the original system, because very good techs which are not prereqs to anything are rarely chosen. I've attempted to compensate for this; but more work is required.

Ultimately I think the AI will also need an overall plan which adds value to certain check-point techs regardless of what is next. (e.g. military techs which will not be obsolete for a long time should be worth more than military techs which have a shorter shelf-life.)

Commit comment from K-Mod git repo

AC4-129