Update unit AI_heal routine
The original code did not take into account the damage param if the unit was not grouped so I have added this so the unit will not attempt to heal if the unit has not breached this threshold.

Added a fort attack check for solo units so they do not end up in a heal/take damage cycle. If the unit is in a plot where it is subject to a fort attack it will not attempt to heal. Trying to find a safe spot could end up with the unit getting stuck so we just abort.

Removed an unused variable.

AC4-166