Add Hurry automation
Added a unit automation to hurry the construction of a building in an available city. The unit prioritises cities with a population of > 2/3 the player average, if there are no cities that match the population criteria then all cities are candidates. From the candidates the unit selects the one that will benefit the most given the time to get to the city.

This automation is not available to GPs

AC4-170