Slaver units get correct promotion availablilty
The code to check if a promotion was valid for a unit looked at the UnitInfo object to determine the UnitCombatType. For slavers this does not work as we change the value as part of the process of becoming a slaver and it is stored as part of the unit object.

The code to check for promotion validity has been updated to allow for checks on either the unit or the base unitinfo. There are some scenarios, such as in the CityAI where there is no unit to check against and in those situations the original code is still used.

The PlatyUI world builder promotion screen has also been updated to use the new functions.

AC4-122