Make slavers slave capacity based on unit strength
The base number of slaves that a slaver can carry was set at 1 for all units. This change makes it so that a unit can manage 1 slave for every 8 strength or part thereof.

The actual calculation is based on the float value of the new strength of the slaver so it is possible for a slaver to have a final strength of 8 with 2 slave slots. The Musketman is an example of this where their base strength of 9 converts to 8.1 after it is modified giving 2 slave slots as that value rounds up, but only 8 combat strength as that value rounds down.

AC4-122