Remove enslavement capability from unit infos
Now that any unit type may become a slaver and the number of slaves is set based on the unit combat strength, the enslavement value in the unit info has become obsolete so this has been removed.

The max slaves code at the unit level has been refactored to reflect this.

The AI to decide when to create slavers is no longer based on the base unit so the AI value has been removed from the check in PlayerAI

AC4-122