Add weapon materials
The material that a units weapons are made of impacts the unit strength. Only some, typically early, unit combat types are impacted by weapon material whilst others are unaffected.

The materials are defined in their own XML file and contain details about which unit combats can be affected, any bonus that is required and the additional strength it gives to units wielding it.

The main interface has been updated to show the weapon type a unit it using as an icon before the list of combat types

AC4-196