Organize unit combat infos into functional and hidden subcategories
Unit subcategories are used in multiple ways, some are to introduce additional functionality, like the ability to attach from forts. Others are are used to indicate applicability for other objects, such as promotions or weapon materials.
Those used to for applicability should typically be hidden on the UI as it is the attached object that is the key. To this end if a unit combat has no button it will not be shown in the pedia.

The UnitCombatInfos file has been commented to clarify these types of combat types for convenience.

AC4-196