Add projectile materials
For weapons that use projectiles of some kind, whether they are arrows or bullets, the type of projectile can impact how effective the weapon is. A new XML type should be created that defines ammunition types based on available city resources, the UnitCombatType of a unit defines which ammunition type is applicable to units.

A unit can also be limited to the maximum bonus received from ammunition by defining tiers for the projectile types and having a value, defined at the unit level of the maximum tier level they can use.

AC4-196