Replace palace as starting building
The palace building is too powerful for the stoneage eras and needs to be replaced.

Created a new starting building that represents the capital in the stoneage automatically upgrades to the palace on entering the Ancient Era

New building is named 'Capital Seat' and provides 1 happiness and 1 production as it's only yield.

AC4-200