Refactor property to determine if leaders get traits
The Tech XML property that was used to determine when leaders get traits was named to be specific to this function, but the tech has wider usage and so property has been renamed to a more generic name to reflect this. The CvTeam and other object properties have also been updated to match.

AC4-203