Allow more XP from animals and barbarians
In the start of the game most combat is from animals and barbarians and in order for the early units to gain promotions it is necessary for them to be able to gain XP from these enemies. This commit increases the max XP gained from animals and barbarians up to the point that one civ becomes settled at that point all civs have the max XP gained capped to the original value. This is to try and prevent one civ from farming XP before moving to settled status.

AC4-206