Add salvage promotion to field promotions and update chance
Added the new Salvage promotion to the list of those available as field promotions.

Simplified the calculation to get the chance of receiving as field promotion and reduced the chance. The original algorithm was giving a chance of a promotion of up to 100%; the new algorithm reduces this to around 10%

AC4-205
AC4-96