Improve AI Healing routine
Healing solo units ignores the damage threshold as it only pushes the mission in cities or if the heal only takes a single turn.

Fixed bug in group healing where the damage threshold param was set to the min of 10 and the param value rather than the maximum as intended.

AC4-166