Give Scout UNITCOMBAT_HUNTER as a sub combat type
Scouts are available prior to  civ being settled and being an upgrade to the Atlatl needs to be able to perform its job. It does not have the fixed AI of the Atlatl, but should be able to retain its scavenging type promotions.

It also has the hunter unit AI as a default, but retains the explorer as well.

AC4-205