Prevent features from overwriting natural wonders
The Natural Wonders are placed on the map before normal features to allow multi-plot wonders to find a viable position more easily. The normal features being placed afterwards did not check whether there was a NW on a plot so could overwrite it. They also did not remove the extra plot yields that come with the NW. This has been fixed so that a plot is not considered valid for a feature if there is a pre-existing feature there.

AC4-102