Limit barbarian units in the early game
There should be no human barbarian units in the Epipalaeolithic era as no civs are considered to have much of an identity so there are no marauding humans; the danger being from animals. Once the Mesolithic era arrives there will be some human barbarians, but they will not be developed enough for cities.

This does not prevent the civ AIs from converting their units to barbarians if they meet the prereqs.

AC4-210