Prevent unique features being removed when normalising starting plots
The routine that removed bad features from around the civ starting plots treated natural wonders as bad as they have no inherent yield and removed them accordingly. This did not remove the extra yield produced when the feature was initially placed leaving a plot with an incorrectly high yield.

This update changes the behaviour to leave unique features in place.

AC4-102