Add Barbarian Leader unit
Added the BL unit, which is a Great Person type. The unit is hidden nationality, always hostile and invisible.

There are 2 available missions and the AI can also use these, thought the logic around which to take needs some work as it is very crude.

1 - Increase the number of free military units that a civ can maintain.
2 - Plunder a rival civ city.

The free units concept needs a dampener so that it doesn't become overpowered. I track the barbarian free unit value separately form the standard one so that I can add some form of decay mechanism in future without impacting other form of this bonus.

Plundering a city takes food, resources and population from the target city to the civ capital.

Future updates may modify the missions, but I wanted to get the concept in there first and these are the missions included in the Stone Age mod that this is taken from.

In order to receive a BL typically you need to win combat using a barbarian unit, but I have not created those yet so this concept should be considered Alpha code.

AC4-209