Add ability to show Great General and Barbarian Leader progress bars
Currently only the Great Person progress is displayed on the main screen. Added functionality to also show the GG and BL progress. These bars are dynamically handled so that none, 1 or more are able to be displayed and the size of each bar alter based on the number shown.

The options dialog has values that can be set to scale each progress bar relative to the others. A progress bar is only be displayed if there is a non-zero value for the progress.

A new font icon has bee created for the BL and added to the GameFonts files. I have been unable to get the larger game font icon to display, only the one from the GameFont75 file. This has the effect that if the font size when displaying this on the progress bar is increased to 3 the icon does not display

AC4-215