Prevent cities spamming Warriors to protect against animals
At the stage in the game when Warriors are available the biggest danger is from animals, but these do not enter a civs borders so are a minimal threat. The city AI did not recognise this so would spam Warriors when other options may be better.

This experimental change modifies the concept of danger to exclude animal threats when deciding city production targets,

AC4-216