Update java tools to correctly order Era Infos and create xml folder structure
The java tooling on import would require that the era sheet in the workbook was sorted by the Advanced Start Cost. If this was forgotten then the Eras in the game would also be out of order. This change corrects that so that the xml file created will automatically order the eras in the file correctly.

The import operation also now creates the standard xml folder structure to make it easier to diff the generated xml files and copy them to the mod folder.

AC4-186