Add additional checks for natural wonders when normalising starting plots
There were still plots where it appears that a natural wonder has been removed, leaving the extra plot yields. These appear to be where a different feature has been placed there instead. The likeliest candidate is the code to increase the city production as I don't recall seeing any where there was not a forest on the plot. I have updated this code to check for a NW before forcing a forest. Hopefully this will finally resolve the issue.

AC4-102