Prevent spawning improvements filling areas
If a spawning improvement is on a small island it may fill all the plots and prevent any player landing there. To prevent this spawning improvements should not be placed on small islands with 2 of fewer tiles and should not spawn more units if there are no free unowned tiles.

AC4-223