Rework forests and forest improvements
In the stone age civs should have limited resource production capabilities, yet a forested hill with a timber camp would be producing 3 hammers, which is out of proportion. It would be possible to cap the hammer yield limit to 2, as happens with food and commerce, but this would remove the value of a timber camp improvement on these plots so instead the hammer yield of forests is removed. This means that a timber camp is required to get the production bonus until the advent of Advanced Tools which gives forests their normal yield back.

The 3 forest improvements have also been tweaked:

- Timber Camp icon updated, upgrades to Lumberjack Camp
- Lumberjack Camp icon updated and gives 1 extra hammer with Railroads, plus 1 commerce when beside a river
- Lumbermill no longer has a river commerce, but has both production and commerce from Railroads

AC4-224