Rework Seed Camp & Stone Tools Camp
The Seed Camp has had it's commerce removed and upgrades to a farm as in the stone age commerce should be very hard to come by and techs were popping too easily. The Seed camp upgrades to a Farm as a natural progression as technologies improve.

The Stone Tools camp upgrades to a Mine for the same reason

AC4-224