Prevent range bound units from grouping with better ranged units
When units group the effective range type of the group is that of the most restrictive range limit. In the early game especially when there are unit types with very restrictive range types this can prevent some units from performing their function. Tribes may be prevented from leaving their territory to settle or barbarians being able to reach other civs. This change should resolves this type of issue.

AC4-41