Add tech prereq to Natural Wonder buildings
The Natural Wonder buildings were too powerful in the stone age, especially the culture boost. The nature yield bonus they provide is sufficient boost at that time. Each of the buildings has a prereq in the first tier of the ancient era and one that is appropriate to the bonuses applied.

AC4-102