Improve value of Barbarian Lodge national wonder
For a national wonder, the Barbarian Lodge was very underwhelming at 2 espionage. Espionage is of very little use in the stone age. This change addresses this in 2 ways, firstly the espionage bonus is replaced with a boost to commerce that can be spent wherever is best, it also adds a thematic change to remove the additional cost for civ barbarian units making them cost the same as standard units.

AC4-209