Allow building to provide a production bonus to other building classes
The construction of one building should be able to provide a production bonus to buildings of another building class. This bonus is a percentage change to the hammers the city has as long as those hammers are being put towards the building class construction and the value should be rounded down. This means that in the very early game when a city has very few hammers there may not be a visible change as the bonus may not equate to a full hammer.

AC4-232