Fixed incorrect child tag in BonusInfo.FeatureTerrainBooleans
The child tag was set as FeatureType when it should have been TerrainType. This did not impact the dll processing as it is set to optional and the code read the children as an unstructured array so didn't check the tag name. The java code generator however does us all child tag names so was producing invalid java code

AC4-11