Rework building vicinity bonus yield changes being active if the bonus is not connected
The XML tag that allowed a building to provide a yield change to a city if it had a bonus in the vicinity, irrespective of whether the bonus is connected or not has been rework to reverse the logic of the XML tag. The tag now needs to be set to true to allow the effect which is in line with the way other tags work. Previously setting this to true would prevent the functionality.

AC4-233