Add Great Jester unit
Added the unit, settled specialist and associated missions.

The GJ unique mission if to perform a Great Jest that will increase the happiness of all player cities in the same area for a set duration. Any city that the GJ is present in will have no unhappiness and any city with an occupation timer will have the timer reduced by an addition 1 every turn that the GJ is present in the city at the end of its turn.

AC4-245
AC4-246