Spawning improvements create resident based on existing plot animals
If the resident of an animal spawning improvement was killed then a new resident was created using a random animal type that fit the terrain. This meant that there could be multiple animal types stacked on the improvement. his often happened if the improvement was on a small plot of land that was bound by water and player territory and the player killed the resident, but not all other animals on the plot.
This change now looks to see if there is an existing animal type on the plot and if so uses that as the resident type.

AC4-188