Fix bug where yield was not being provided to city with vicinity bonus if unconnected
The building flag to indicate that a vicinity bonus should provide a yield to a city even if it is not connected to the city trade network was not being respected. It was working when determining if a building could be built, but not for the yield. This commit fixes this.

AC4-231