Create un-researchable tech
An un-researchable tech can be used as a prereq for other resources such as Promotions allowing them to be prohibited from selection under normal circumstances and only available for direct selection via the game engine. An example of this may be a promotion that should only be given to a unit in specific circumstances such as a city being besieged or other situation where the creation of a prereq tag for the scenario would not be worth the additional memory usage and complexity.

AC4-260