Do not alpha-sort Civics in Pedia 'Type' view
In the Pedia it is possible to show the civics using multiple sorting styles, one of these is to sort them by the CivicOptionType. This sort currently has the civics within each option type alpha sorted but it is more useful to show them in the order they appear in the XML file as this allows the order that they are available to be used instead, assuming that the XML file is created that way.

AC4-264