Update and add add new Civics
The civic imported are essentially those used in RoM:AND

The new civics have been created to allow their reference in buildings and traits, etc., but they require rebalancing and reworking to simplify them as some have too many effects and especially too many negative modifiers. The civics should be adding to the game play and not reducing it as a general principal.

Java toolset updated accordingly

AC4-236
Update and add add new Civics

The civic imported are essentially those used in RoM:AND

The new civics have been created to allow their reference in buildings and traits, etc., but they require rebalancing and reworking to simplify them as some have too many effects and especially too many negative modifiers. The civics should be adding to the game play and not reducing it as a general principal.

AC4-236