Increase unit rehoming chance with length of stay in city
The chance that a unit is rehomed to a new city should increase with number of consecutive times it ends its turn in the target city. This commit multiplies the base chance  by the number of turns ended in the city. If the base chance is 5% then by the second turn in the city this will increase to 10%, then 15%, and so on.

AC4-289